pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "splitter-bench-rag"
version = "0.1.0"
description = "End-to-end benchmark for LangChain text splitters using pgvector and OpenAI"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  # Core
  "python-dotenv>=1.0.1",
  "pydantic>=2.6.0",
  # LangChain
  "langchain>=0.2.14",
  "langchain-core>=0.2.35",
  "langchain-community>=0.2.12",
  "langchain-text-splitters>=0.3.2",
  "langchain-postgres>=0.0.11",
  # Vector DB
  "pgvector>=0.3.2",
  "psycopg[binary]>=3.1.18",
  "sqlalchemy>=2.0.25",
  # OpenAI
  "openai>=1.40.0",
  "tiktoken>=0.7.0",
  # Data
  "numpy>=1.26.4",
  "pandas>=2.2.2",
  "pyyaml>=6.0.1",
  # PDF + HTML
  "pymupdf>=1.24.3",
  "beautifulsoup4>=4.12.3",
  "markdown>=3.6",
  "lxml>=5.2.1",
  # Plotting
  "matplotlib>=3.8.4",
  # Utils
  "tqdm>=4.66.4",
  "rich>=13.7.1",
  "requests>=2.32.3",
  "langchain-openai>=1.1.7",
  "tabulate>=0.9.0",
]

[tool.setuptools.packages.find]
where = ["src"]
